// project4type-redux/ecommercetypeproj4/src/models/Userapi.ts
export interface IUser {
    // id: number;
    // name: string;
    // // username: string;
    // email: string;


    id: number;
    name: string;
    title: string;
    price: number;
    description: string;
    category: string;
    image: string;
}

// id: number;
// name: string;
// title: string;
// price: number;
// description: string;
// category: string;
// image: string;